+++ sdk/nodejs/config/vars.ts
@@ -7,17 +7,23 @@ import * as utilities from "../utilities";
 let __config = new pulumi.Config("openstack");
 
 /**
+ * If set to `true`, OpenStack authorization will be perfomed automatically, if the initial auth token get expired. This is
+ * useful, when the token TTL is low or the overall Terraform provider execution time expected to be greater than the
+ * initial token TTL.
+ */
+export let allowReauth: boolean | undefined = __config.getObject<boolean>("allowReauth") || utilities.getEnvBoolean("OS_ALLOW_REAUTH");
+/**
  * Application Credential ID to login with.
  */
-export let applicationCredentialId: string | undefined = __config.get("applicationCredentialId");
+export let applicationCredentialId: string | undefined = __config.get("applicationCredentialId") || utilities.getEnv("OS_APPLICATION_CREDENTIAL_ID");
 /**
  * Application Credential name to login with.
  */
-export let applicationCredentialName: string | undefined = __config.get("applicationCredentialName");
+export let applicationCredentialName: string | undefined = __config.get("applicationCredentialName") || utilities.getEnv("OS_APPLICATION_CREDENTIAL_NAME");
 /**
  * Application Credential secret to login with.
  */
-export let applicationCredentialSecret: string | undefined = __config.get("applicationCredentialSecret");
+export let applicationCredentialSecret: string | undefined = __config.get("applicationCredentialSecret") || utilities.getEnv("OS_APPLICATION_CREDENTIAL_SECRET");
 /**
  * The Identity authentication URL.
  */
@@ -41,7 +47,7 @@ export let defaultDomain: string | undefined = __config.get("defaultDomain") || 
 /**
  * If set to `true`, OpenStack authorization will be perfomed, when the service provider client is called.
  */
-export let delayedAuth: boolean | undefined = __config.getObject<boolean>("delayedAuth");
+export let delayedAuth: boolean | undefined = __config.getObject<boolean>("delayedAuth") || utilities.getEnvBoolean("OS_DELAYED_AUTH");
 /**
  * If set to `true`, the HTTP `Cache-Control: no-cache` header will not be added by default to all API requests.
  */