+++ CHANGELOG.md
@@ -4,6 +4,9 @@ CHANGELOG
 ## HEAD (Unreleased)
 * Namespace names in .NET SDK are adjusted to PascalCase
 ([#70](https://github.com/pulumi/pulumi-openstack/pull/70)).
+* Upgrade to v1.25.0 of the OpenStack Terraform Provider
+* Upgrade to pulumi-terraform-bridge v1.5.2
+
 ---
 
 ## 1.2.0 (2019-12-04)